ENH: Organizing the code better and adding better CLI help.
The code is now better organized because the NBIA client functionality is
broken out into its own class. This leaves the main method of the main class
to do nothing but argument parsing and config loading before handing off the
necessary data to the NBIA simple client. The AutoHelpParser is created to
automate the creation of help output.